Initialized registration metric benchmark. Moved tests in test_registration to registration_module test.
